+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### ✅ Added
 - Make it possible to customize video asset (e.g. include custom HTTP header) before it's preview/content is loaded [#1510](https://github.com/GetStream/stream-chat-swift/pull/1510)
 - Make it possible to search for messages containing attachments of the given types [#1525](https://github.com/GetStream/stream-chat-swift/pull/1525)
+- Make `ChatReactionsBubbleView` open for UI customization [#1526](https://github.com/GetStream/stream-chat-swift/pull/1526)
 
 ### 🔄 Changed
 - Rename `VideoPreviewLoader` type to `VideoLoading` and `videoPreviewLoader` to `videoLoader` in `Components` [#1510](https://github.com/GetStream/stream-chat-swift/pull/1510)